thea_manager_backend/src_2/requirements.txt
boto3
flask
typeguard
Flask-Cors
flask_bcrypt
python-dotenv
PyYAML
pyjwt
python-jose
requests